refactor(test): split theme appearance tests by builtin example override and classification behavior
